requirements.txt
transformers==4.44.2
langchain==0.3.0
langchain-openai==0.2.0
datasets[audio]==3.0.0
torchaudio==2.4.1
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.2
python-dotenv==1.0.1
